Return last net.IPNet/net.IP when looking for an net.IP
It seems that returning the last entry of the ContainingNetworks
contains the most specific match when looking for an net.IP, so
we return the slice of entries (values) in the reverse order.
I haven't verified this to be 100% true yet, but so far it looks
OK to do it like this.